Added Basic Menu with CRT Lines, and Screens
Added a basic menu with text displaying the name of the game and a button that has hover animations. Added a screen class with update and render functions to organize the project more. Each class that is a screen inherits from the base Screen class and the main.cpp file handles all calls to update, render, and do other things to make the screens function.